Enhancement: UserGroups module to include bool for has_binding (#2)
UPDATE: Modules::UserGroup to have an option to dynamically determine if a user_group has a role_binding when called from a higher level template.
UPDATE: README documentation
UPDATE: Terraform supported versions
UPDATE: .gitignore to ignore .terraform.lock.hcl files for testing

This update will resolve issues where a template creates a role and resource group during the same execution as the group.  In some cases, it is nice
to have a solution that will optionally add bindings only when those bindings are required as in if one of many groups are created but have no binding
details.